restructure zod schema files
